Heavily refactored Redux codebase
Modularized actions and reducers
