+++ static/css/styles.scss
@@ -30,6 +30,10 @@ input {
     }
 }
 
+textarea {
+    resize: vertical;
+}
+
 .btn,
 .nav-tabs > li > a {
     -webkit-border-radius: 0;
