+++ static/js/src/actions/app.js
@@ -1,0 +1,7 @@
+export const SELECT_TAB = 'SELECT_TAB'
+export function selectTab(tab) {
+    return {
+        type: SELECT_TAB,
+        tab
+    }
+}
