+++ static/js/src/actions/index.js
@@ -1,0 +1,3 @@
+export * from './app.js'
+export * from './notes.js'
+export * from './todos.js'
