+++ static/js/src/actions/notes.js
@@ -1,0 +1,4 @@
+export const CREATE_NOTE = 'CREATE_NOTE'
+export function createNote(data) {
+    debugger
+}
