+++ static/js/src/components/Notes.js
@@ -1,6 +1,13 @@
 import React, { Component } from 'react'
 
 export class Notes extends Component {
+
+    handleSave() {
+        const { createNote } = this.props.actions
+
+        createNote({ hello: 'world' })
+    }
+
     render() {
         if ( ! this.props.isActive) {
             return null
@@ -11,7 +18,7 @@ export class Notes extends Component {
                 <input type="text" className="large-input" placeholder="Title"/>
                 <textarea name="note" id="" cols="30" rows="4" className="large-input"
                     placeholder="Take Note"></textarea>
-                <button className="btn btn-primary btn-lg">Save</button>
+                <button className="btn btn-primary btn-lg" onClick={ this.handleSave.bind(this) }>Save</button>
             </div>
         ) 
     }
