+++ static/js/src/reducers/todos.js
@@ -1,0 +1,5 @@
+const init = {}
+
+export default function todos(state = init, action) {
+    return state
+}
